Transfer checkNetflix to the BrowserWindow class
